+++ src/reporter.ts
@@ -91,7 +91,10 @@ export class RPReporter implements Reporter {
   isLaunchFinishSend: boolean;
 
   constructor(config: ReportPortalConfig) {
-    this.config = config;
+    this.config = {
+      ...config,
+      launchId: process.env.RP_LAUNCH_ID || config.launchId,
+    };
     this.suites = new Map();
     this.suitesInfo = new Map();
     this.testItems = new Map();
@@ -262,7 +265,8 @@ export class RPReporter implements Reporter {
   }
 
   onBegin(): void {
-    const { launch, description, attributes, skippedIssue, rerun, rerunOf, mode } = this.config;
+    const { launch, description, attributes, skippedIssue, rerun, rerunOf, mode, launchId } =
+      this.config;
     const systemAttributes: Attribute[] = getSystemAttributes(skippedIssue);
 
     const startLaunchObj: StartLaunchObjType = {
@@ -274,6 +278,7 @@ export class RPReporter implements Reporter {
       rerun,
       rerunOf,
       mode: mode || LAUNCH_MODES.DEFAULT,
+      id: launchId,
     };
     const { tempId, promise } = this.client.startLaunch(startLaunchObj);
     this.addRequestToPromisesQueue(promise, 'Failed to start launch.');
@@ -565,12 +570,16 @@ export class RPReporter implements Reporter {
       });
       this.finishSuites();
     }
-    const { promise } = this.client.finishLaunch(this.launchId, {
-      endTime: this.client.helpers.now(),
-      ...(this.customLaunchStatus && { status: this.customLaunchStatus }),
-    });
+
+    if (!this.config.launchId) {
+      const { promise } = this.client.finishLaunch(this.launchId, {
+        endTime: this.client.helpers.now(),
+        ...(this.customLaunchStatus && { status: this.customLaunchStatus }),
+      });
+      this.addRequestToPromisesQueue(promise, 'Failed to finish launch.');
+    }
+
     this.isLaunchFinishSend = true;
-    this.addRequestToPromisesQueue(promise, 'Failed to finish launch.');
     await Promise.all(this.promises);
     this.launchId = null;
   }